voice_analytics/retell_integration_sample/requirements.txt
httpx==0.27.0
python-dotenv==1.0.1
fastapi==0.115.6
uvicorn==0.34.0
retell-sdk==4.11.1